ai-service/requirements.txt
torch==2.0.1
torchvision==0.15.2
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
requests==2.31.0